requirements.txt
openai>=1.0.0
slack_sdk>=3.0.0
feedparser>=6.0.0
requests>=2.28.0
apify-client>=1.0.0
python-dotenv>=1.0.0
hypothesis>=6.0.0
pytest>=7.0.0
pyyaml>=6.0.0
